refactor: replace App component files to the 'components' folder
